Fix naming, package location
