+++ build-logic/plugins/android/src/main/kotlin/ru/astrainteractive/gradleplugin/model/AndroidSdkInfo.kt
