+++ build-logic/plugins/android/src/main/kotlin/ru/astrainteractive/gradleplugin/property/extension/AndroidModelPropertyValueExt.kt
@@ -1,0 +1,15 @@
+package ru.astrainteractive.gradleplugin.property.extension
+
+import org.gradle.api.Project
+import ru.astrainteractive.gradleplugin.model.AndroidSdkInfo
+import ru.astrainteractive.gradleplugin.property.PropertyValue.Companion.baseGradleProperty
+import ru.astrainteractive.gradleplugin.property.extension.PrimitivePropertyValueExt.requireInt
+
+object AndroidModelPropertyValueExt {
+    val Project.requireAndroidSdkInfo: AndroidSdkInfo
+        get() = AndroidSdkInfo(
+            compile = baseGradleProperty("android.sdk.compile").requireInt,
+            min = baseGradleProperty("android.sdk.min").requireInt,
+            target = baseGradleProperty("android.sdk.target").requireInt
+        )
+}
