+++ build-logic/plugins/android/src/main/kotlin/ru/astrainteractive/gradleplugin/property/extension/ExtendedAndroidPropertyExt.kt
@@ -4,7 +4,7 @@ import org.gradle.api.Project
 import ru.astrainteractive.gradleplugin.property.PropertyValue.Companion.baseGradleProperty
 import ru.astrainteractive.gradleplugin.property.extension.PrimitivePropertyValueExt.requireInt
 
-object ExtendedProjectPropertyExt {
+object ExtendedAndroidPropertyExt {
     val Project.requireVersionCode: Int
         get() = baseGradleProperty("project.version.code").requireInt
 }
